+++ Sources/flag_x.c
@@ -6,7 +6,7 @@
 /*   By: ugotheveny <ugotheveny@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/25 01:23:30 by ugotheveny        #+#    #+#             */
-/*   Updated: 2020/06/25 01:23:43 by ugotheveny       ###   ########.fr       */
+/*   Updated: 2020/07/07 13:43:24 by ugtheven         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	flag_x(va_list *args)
 {
-	unsigned int n;
+	unsigned long long n;
 
-	n = va_arg(*args, unsigned int);
+	n = va_arg(*args, unsigned long long);
 	ft_puthlow(n);
 }
